Fields for simulator class
