+++ bin/.gitignore
@@ -1,8 +1,4 @@
-/Customer.class
-/Familysedans.class
-/Motorbikes.class
-/Simulator.class
-/SmallCar.class
-/Vehicle.class
-/View/
 /Model/
+/Simulator.class
+/View/
+/aston/
