+++ src/Simulator.java
@@ -1,5 +1,15 @@
+import aston.group17.model.*;
+import aston.group17.view.*;
+
+import java.util.ArrayList;
 
 public class Simulator {
+	private int probabilityP;
+	private int probabilityQ;
+	private int probabilityTruck;
+	
+	private ArrayList<Vehicle> vehicles;
+	
 	public static void main(String[] args)
 	{
