+++ src/aston/group17/model/Customer.java
@@ -1,4 +1,4 @@
-package Model;
+package aston.group17.model;
 
 public class Customer {
 	//test
