+++ src/aston/group17/model/SmallCar.java
@@ -1,26 +1,31 @@
-package Model;
+package aston.group17.model;
 import java.util.Random;
 
-public class Motorbikes implements Vehicle {
+
+public class SmallCar implements Vehicle {
 
 	private int gallons;
 	private double unitSpace;
 	private Random rand;
-
-	public Motorbikes(){
-		gallons =5;
-		unitSpace=0.75;
-		rand = new Random();
 	
+	
+	public SmallCar(){
+		gallons = 7;
+		unitSpace = 1;
+		rand = new Random();
 	}
+	
+	
 	
 	public void setGallons(int gallons) {
-		this.gallons = gallons;
+		int a = rand.nextInt(2 + 1 - 0);
+		gallons = gallons + a;
+		
 	}
 
 
-	/*	public void setUnitSpace(double unitSpace) {
-	
+/*	public void setUnitSpace(double unitSpace) {
+		
 	} */
