+++ src/aston/group17/view/SimulatorGUI.java
@@ -1,4 +1,4 @@
-package View;
+package aston.group17.view;
 
 public class SimulatorGUI {
